+++ src/ui_event/dialog_ui.py
@@ -1,7 +1,6 @@
 import os
 import sys
 
-
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 import sys
@@ -221,6 +220,17 @@ class Dialog(QDialog):
         save_button.clicked.connect(lambda: save_data(self))
         cancel_button.clicked.connect(self.reject)
 
+    def reject(self):
+        """
+
+        :return:
+        """
+        # 在这里你可以添加任何你需要在对话框关闭时执行的代码
+        logger.debug('config msg dialog Dialog is cancel closing!')
+        constants.edit_config_msg_visible = False
+        # 调用基类的 closeEvent 方法以确保对话框正常关闭
+        super(Dialog, self).reject()
+
     def closeEvent(self, event):
         """
         对话框关闭
@@ -228,7 +238,7 @@ class Dialog(QDialog):
         :return:
         """
         # 在这里你可以添加任何你需要在对话框关闭时执行的代码
-        logger.debug('config msg dialog Dialog is closing!')
+        logger.debug('config msg dialog Dialog is close closing!')
         constants.edit_config_msg_visible = False
         # 调用基类的 closeEvent 方法以确保对话框正常关闭
         super(Dialog, self).closeEvent(event)